+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Changed
+
+- Way to enqueue tasks returning if it was successful
+
 ## [0.2.0] - 2019-10-10
 
 ## Added