+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+
+- `canEditData` field to the data provided by `order-shipping`.
+
 ## [0.2.1] - 2019-11-19
 
 ### Changed